Fixed pre-render

+++ packages/cloud/src/index.tsx
@@ -10,10 +10,14 @@ console.log('Version: ' + process.env.npm_package_version);
     // Handle ?= search headers
     let a = handleSearch(location.search, location.protocol + "//" + location.host + location.pathname)
     a.length > 0 && location.replace(a)
+    if (a.length > 0)
+        return;
 
     // Handle Redirecting if the user isnt logged in
     let b = checkToken(localStorage.getItem('token'), 'https://auth.lvk.sh/login?callback=' + location.toString())
     b.length > 0 && [localStorage.clear(), location.replace(b)]
+    if (a.length > 0)
+        return;
     
     // Render root
     var root = document.getElementById("root")
